System.Console.Pretty

Description

Useful helpers to style and color text with ANSI escape sequences.

Documentation

data Color

Colors for an ANSI terminal

Constructors

Black

Red

Green

Yellow

Blue

Magenta

Cyan

White

Default

class Pretty a where

A class to color and style

Minimal complete definition

colorize, style

Methods

color :: Color -> a -> a

Helper to set foreground color

bgColor :: Color -> a -> a

Helper to set background color

colorize :: Section -> Color -> a -> a

Set the color of the (fg/bg) with the color

style :: Style -> a -> a

Set the style

data Section

A section to be colored, either foreground or background.

Constructors

Foreground

Background

data Style

SGR paramaters, aka text styles for an ANSI terminal

Constructors

Normal

Bold

Faint

Italic

Underline

SlowBlink

ColoredNormal

Reverse

supportsPretty :: IO Bool

Whether or not the current terminal supports pretty-terminal